// src/app/model/master/syllabus-subject-details.model.ts
export class SyllabusSubjectDetails {
  subjectDetailsId: number;
  subject: string;
  fa1: string;
  fa2: string;
  sa1: string;
  fa3: string;
  fa4: string;
  sa2: string;

  constructor() {
    this.subjectDetailsId = 0;
    this.subject = '';
    this.fa1 = '';
    this.fa2 = '';
    this.sa1 = '';
    this.fa3 = '';
    this.fa4 = '';
    this.sa2 = '';
  }
}